Move shell into heroic-dist module.
