+++ heroic-dist/build.gradle
@@ -1,3 +1,4 @@
+apply plugin: 'com.google.protobuf'
 apply plugin: 'application'
 apply plugin: 'com.github.johnrengelman.shadow'
 
@@ -29,10 +30,13 @@ dependencies {
     implementation 'args4j:args4j'
     implementation 'com.google.guava:guava'
     implementation 'eu.toolchain.async:tiny-async-core'
+    implementation 'jline:jline:2.12'
+    implementation 'com.google.protobuf:protobuf-java'
+    
     api project(':heroic-all')
-    api project(':heroic-shell')
     api project(':heroic-core')
     api project(':heroic-component')
+    
     compileOnly 'org.projectlombok:lombok'
 
     testCompileOnly 'org.projectlombok:lombok'
