+++ heroic-dist/src/main/java/com/spotify/heroic/HeroicShell.java
@@ -26,8 +26,10 @@ import com.spotify.heroic.HeroicCore.Builder;
 import com.spotify.heroic.args4j.CmdLine;
 import com.spotify.heroic.proto.ShellMessage.CommandsResponse.CommandDefinition;
 import com.spotify.heroic.shell.CoreInterface;
+import com.spotify.heroic.shell.DirectShellIO;
 import com.spotify.heroic.shell.RemoteCoreInterface;
 import com.spotify.heroic.shell.ShellIO;
+import com.spotify.heroic.shell.ShellParameters;
 import com.spotify.heroic.shell.ShellTask;
 import com.spotify.heroic.shell.TaskParameters;
 import eu.toolchain.async.AsyncFramework;
