+++ heroic-dist/src/main/java/com/spotify/heroic/shell/DirectShellIO.kt
@@ -19,25 +19,29 @@
  * under the License.
  */
 
-package com.spotify.heroic
+package com.spotify.heroic.shell
 
-import com.spotify.heroic.shell.AbstractShellTaskParams
-import java.util.ArrayList
-import org.kohsuke.args4j.Option
+import com.spotify.heroic.shell.ShellIO
+import java.io.IOException
+import java.io.InputStream
+import java.io.OutputStream
+import java.io.PrintWriter
+import java.nio.file.Files
+import java.nio.file.Path
 
-class ShellParameters : AbstractShellTaskParams() {
-    @Option(name = "--server", usage = "Start shell as server (enables listen port)")
-    val server = false
+data class DirectShellIO(val out: PrintWriter?) : ShellIO {
 
-    @Option(name = "--shell-server", usage = "Start shell with shell server (enables remote connections)")
-    val shellServer = false
+    @Throws(IOException::class)
+    override fun newInputStream(path: Path): InputStream {
+        return Files.newInputStream(path)
+    }
 
-    @Option(name = "--disable-backends", usage = "Start core without configuring backends")
-    val disableBackends = false
+    @Throws(IOException::class)
+    override fun newOutputStream(path: Path): OutputStream {
+        return Files.newOutputStream(path)
+    }
 
-    @Option(name = "--connect", usage = "Connect to a remote heroic server", metaVar = "<host>[:<port>]")
-    val connect: String? = null
-
-    @Option(name = "-X", usage = "Define an extra parameter", metaVar = "<key>=<value>")
-    val parameters = ArrayList<String>()
+    override fun out(): PrintWriter? {
+        return out
+    }
 }
