+++ heroic-dist/src/main/java/com/spotify/heroic/shell/QuoteParserException.java
