+++ heroic-dist/src/main/java/com/spotify/heroic/shell/ShellParameters.kt
@@ -19,29 +19,25 @@
  * under the License.
  */
 
-package com.spotify.heroic
+package com.spotify.heroic.shell
 
-import com.spotify.heroic.shell.ShellIO
-import java.io.IOException
-import java.io.InputStream
-import java.io.OutputStream
-import java.io.PrintWriter
-import java.nio.file.Files
-import java.nio.file.Path
+import com.spotify.heroic.shell.AbstractShellTaskParams
+import java.util.ArrayList
+import org.kohsuke.args4j.Option
 
-data class DirectShellIO(val out: PrintWriter?) : ShellIO {
+class ShellParameters : AbstractShellTaskParams() {
+    @Option(name = "--server", usage = "Start shell as server (enables listen port)")
+    val server = false
 
-    @Throws(IOException::class)
-    override fun newInputStream(path: Path): InputStream {
-        return Files.newInputStream(path)
-    }
+    @Option(name = "--shell-server", usage = "Start shell with shell server (enables remote connections)")
+    val shellServer = false
 
-    @Throws(IOException::class)
-    override fun newOutputStream(path: Path): OutputStream {
-        return Files.newOutputStream(path)
-    }
+    @Option(name = "--disable-backends", usage = "Start core without configuring backends")
+    val disableBackends = false
 
-    override fun out(): PrintWriter? {
-        return out
-    }
+    @Option(name = "--connect", usage = "Connect to a remote heroic server", metaVar = "<host>[:<port>]")
+    val connect: String? = null
+
+    @Option(name = "-X", usage = "Define an extra parameter", metaVar = "<key>=<value>")
+    val parameters = ArrayList<String>()
 }
