+++ heroic-dist/src/test/java/com/spotify/heroic/shell/QuoteParserTest.java
