+++ settings.gradle
@@ -7,7 +7,6 @@ include(':heroic-component')
 include(':heroic-elasticsearch-utils')
 include(':heroic-core')
 include(':heroic-test')
-include(':heroic-shell')
 include(':heroic-all')
 include(':heroic-metric-datastax')
 include(':heroic-metric-bigtable')
